fix: improve code quality for Codeberg/Gitea PR
- Add backend/.env.example with GITEA_TOKEN documentation
- Extract shared CommitInfo interface to backend/src/types.ts
- Add proper GiteaCommit type for API response (removes 'as any[]')
- Add error/clearError state to useCodeberg.tsx for OAuth failures
- Replace empty catch blocks in service.ts with console.debug logging

🤖 Authored by Loki
