+++ backend/.env.example
@@ -1,0 +1,10 @@
+# GitHub API token (optional, increases rate limits)
+GITHUB_TOKEN=
+
+# Gitea/Codeberg API token (optional, increases rate limits)
+# For Codeberg: Generate at https://codeberg.org/user/settings/applications
+# For self-hosted Gitea: Generate at https://your-gitea-instance/user/settings/applications
+GITEA_TOKEN=
+
+# Private key for signing attestations
+PRIVATE_KEY=
